pyproject.toml
# CloudOps: Incident Responder Environment
# Copyright (c) 2026 CloudOps Contributors
# Licensed under the MIT License

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudops-openenv"
version = "0.1.0"
description = "CloudOps Incident Responder: High-fidelity SRE benchmark"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]
package-dir = { "server" = "server" }
